Refactoring to split large module
