+++ examples/monitor.rs
@@ -11,9 +11,6 @@
 #[cfg(feature = "tokio02")]
 extern crate tokio_02 as tokio;
 
-#[cfg(feature = "tokio03")]
-extern crate tokio_03 as tokio;
-
 #[cfg(feature = "tokio10")]
 extern crate tokio_10 as tokio;
