+++ examples/psubscribe.rs
@@ -11,9 +11,6 @@
 #[cfg(feature = "tokio02")]
 extern crate tokio_02 as tokio;
 
-#[cfg(feature = "tokio03")]
-extern crate tokio_03 as tokio;
-
 #[cfg(feature = "tokio10")]
 extern crate tokio_10 as tokio;
 
@@ -21,7 +18,7 @@ use std::env;
 
 use futures::StreamExt;
 
-use redis_async::{client, resp::FromResp};
+use redis_async::{client, protocol::FromResp};
 
 #[tokio::main]
 async fn main() {
